fix: relocation and rename itemcard
