+++ src/components/atom/index.jsx
@@ -1,4 +1,5 @@
 import HeaderNav from './Header';
+import ItemCard from './ItemCard';
 import SpinnerScreen from './SpinnerScreen';
 
 export { HeaderNav, SpinnerScreen };
